UGENE-6504. Memory reallocating error when exporting big fasta to CLUSTALW (#494)
* UGENE-7367. Optimize 'Tools->Generate sequence' tool

* Self-review fixes

* Rename reserve -> quota

* Make evaluate method static

* UGENE-6504. Memory reallocating error when exporting big fasta to CLUSTALW

* Ugene 7074 (#493)

* UGENE-7074 Help button in Get amino translation of a sequence scenario points to the wrong page

* UGENE-7074. Help button in Get amino translation of a sequence scenario points to the wrong page

* UGENE-5941. Auto annotations updates don't take into account sequence length changings (#489)

* UGENE-6797. Welcome page should not show links to non-available file or projects (#491)

* UGENE-6797. Welcome page should not show links to non-available files/projects.

* Code review fixes

* Fix flaky GUITest_common_scenarios_msa_editor_test_0007_3

* Fix GUITest_common_scenarios_start_page_test_0005

* Fix for GUITest_regression_scenarios_test_4985

+++ src/corelibs/U2View/src/ov_sequence/DetView.cpp
@@ -626,22 +626,22 @@ void DetView::updateVisibleRange() {
             visibleLinesCount++;
         }
 
-        int visibleRangeLen = visibleLinesCount * detArea->getSymbolsPerLine();
+        qint64 visibleRangeLen = visibleLinesCount * detArea->getSymbolsPerLine();
 
-        int lastLine = seqLen / detArea->getSymbolsPerLine() + (seqLen % detArea->getSymbolsPerLine() == 0 ? 0 : 1) - detArea->getLinesCount();
+        qint64 lastLine = seqLen / detArea->getSymbolsPerLine() + (seqLen % detArea->getSymbolsPerLine() == 0 ? 0 : 1) - detArea->getLinesCount();
         if (detArea->height() - detArea->getLinesCount() * detArea->getShiftsCount() * detArea->getShiftHeight() > 0 && lastLine > 0) {
             lastLine--;
         }
-        int lastStartPos = lastLine * detArea->getSymbolsPerLine();
+        qint64 lastStartPos = lastLine * detArea->getSymbolsPerLine();
 
         bool posAtTheEnd = visibleRange.startPos > lastStartPos;
-        visibleRange.length = qMin((int)(seqLen - visibleRange.startPos), qMin(visibleRangeLen, (int)seqLen));
+        visibleRange.length = qMin(seqLen - visibleRange.startPos, qMin(visibleRangeLen, seqLen));
         bool emptyLineDetected = (visibleRangeLen - visibleRange.length) > detArea->getSymbolsPerLine();
 
         if (posAtTheEnd || (emptyLineDetected && visibleRange.startPos + visibleRangeLen >= seqLen)) {
             // scroll to the end
             visibleRange.startPos = qMax(0, (verticalScrollBar->maximum() / numShiftsInOneLine) * detArea->getSymbolsPerLine());
-            visibleRange.length = qMin((int)(seqLen - visibleRange.startPos), qMin(visibleRangeLen, (int)seqLen));
+            visibleRange.length = qMin(seqLen - visibleRange.startPos, qMin(visibleRangeLen, seqLen));
             currentShiftsCounter = qMax(0, verticalScrollBar->maximum() % numShiftsInOneLine);
         }
     } else {
@@ -690,7 +690,7 @@ void DetView::updateVerticalScrollBar() {
     int maximum;
     if (isWrapMode()) {
         DetViewRenderArea *detArea = getDetViewRenderArea();
-        int linesCount = seqLen / detArea->getSymbolsPerLine();
+        qint64 linesCount = seqLen / detArea->getSymbolsPerLine();
         if (seqLen % detArea->getSymbolsPerLine() != 0) {
             linesCount++;
         }
